--- unknown-note-server/src/main/java/unknownnote/unknownnoteserver/dto/DiaryDTO.java
@@ -1,15 +1,0 @@
-package unknownnote.unknownnoteserver.dto;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
-
-@Data
-@RequiredArgsConstructor
-public class DiaryDTO {
-
-    private String dcontent;
-    private String dtag;
-    private int openable;
-
-
-}
