--- unknown-note-server/src/main/java/unknownnote/unknownnoteserver/dto/UserInfo.java
@@ -1,14 +1,0 @@
-package unknownnote.unknownnoteserver.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
-@AllArgsConstructor
-public class UserInfo {
-    private int user_id;
-    private String nickname;
-    private String introduction;
-}
