--- unknown-note-server/src/main/resources/application.properties
@@ -1,19 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/diary_db?useSSL=false&characterEncoding=UTF-8&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=9912
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-
-server.port=8080
-
-spring.datasource.hikari.maximum-pool-size=5
-spring.datasource.hikari.minimum-idle=1
-spring.datasource.hikari.idle-timeout=60000
-spring.datasource.hikari.pool-name=SpringBootJPAHikariCP
-
-#jwt
-jwt.secret = ilovekwangwoonuniversity2019203094
-
