--- unknown-note-server/src/test/java/unknownnote/unknownnoteserver/UnknownNoteServerApplicationTests.java
@@ -1,13 +1,0 @@
-package unknownnote.unknownnoteserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UnknownnoteserverApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
