+++ unknownnoteserver/entity/Diary.java
@@ -19,18 +19,13 @@ public class Diary {
 
     @Column(name = "d_content")
     private String dContent;
-
     @Column(name = "d_time")
     private Timestamp diaryTime;
-
     @Column(name = "d_tag")
     private String dTag;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id", referencedColumnName = "user_id", nullable = false)
-    private User user;
-
-    @Column(name = "openable", nullable = false)
+    @Column(name = "user_id")
+    private int userId;
+    @Column(name = "openable")
     private int openable;
 
 }
