+++ unknownnoteserver/entity/User.java
@@ -1,0 +1,18 @@
+package unknownnote.unknownnoteserver.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "user")
+@Getter
+@Setter
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int user_id;
+    private String nickname;
+    private String introduction;
+
+}
